- Add question, quizz exception handling and refactor QuizzController,...

See merge request hcm24_cpl_java_12/team-2/cursus!37

+++ src/main/java/com/example/Mini_Project1/service/QuizzService.java
@@ -1,26 +1,30 @@
 package com.example.Mini_Project1.service;
 
-import com.example.Mini_Project1.entity.Chapter;
-import com.example.Mini_Project1.entity.Course;
-import com.example.Mini_Project1.entity.Quizz;
-import com.example.Mini_Project1.repository.ChapterRepository;
-import com.example.Mini_Project1.repository.CourseRepository;
-import com.example.Mini_Project1.repository.QuizzRepository;
-import com.example.Mini_Project1.request.QuizzAndQuestion.*;
-import com.example.Mini_Project1.response.QuizzAndQuestion.QuizzResponse;
-import jakarta.transaction.Transactional;
-import lombok.AllArgsConstructor;
-import lombok.RequiredArgsConstructor;
-import org.modelmapper.TypeToken;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.stereotype.Service;
-
 import java.util.Date;
 import java.util.List;
 import java.util.UUID;
 
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.stereotype.Service;
+
+import com.example.Mini_Project1.entity.Chapter;
+import com.example.Mini_Project1.entity.Course;
+import com.example.Mini_Project1.entity.Quizz;
+import com.example.Mini_Project1.exception.AccessDeniedException;
+import com.example.Mini_Project1.exception.BadRequestException;
+import com.example.Mini_Project1.exception.NotFoundException;
+import com.example.Mini_Project1.repository.ChapterRepository;
+import com.example.Mini_Project1.repository.CourseRepository;
+import com.example.Mini_Project1.repository.QuizzRepository;
+import com.example.Mini_Project1.request.QuizzAndQuestion.CreateQuizzByChapterRequest;
+import com.example.Mini_Project1.request.QuizzAndQuestion.CreateQuizzByCourseRequest;
+import com.example.Mini_Project1.request.QuizzAndQuestion.UpdateQuizzRequest;
+import com.example.Mini_Project1.response.QuizzAndQuestion.QuizzResponse;
+import com.nimbusds.jose.shaded.gson.reflect.TypeToken;
+
+import jakarta.transaction.Transactional;
+import lombok.AllArgsConstructor;
 
 @Service
 @AllArgsConstructor
@@ -29,24 +33,25 @@ public class QuizzService {
     private final QuizzRepository quizzRepository;
     private final CourseRepository courseRepository;
     private final ChapterRepository chapterRepository;
+    private final ModelMapper modelMapper;
 
     @Transactional
-    // create quizz by course id
-    public QuizzResponse createQuizzByCourseService(CreateQuizzByCourseRequest request) {
+    public QuizzResponse createQuizzByCourseService(CreateQuizzByCourseRequest request, UserDetails userDetails) {
         Course course = courseRepository.findById(request.getCourseId().toString()).orElseThrow(
-                () -> new RuntimeException("Course not found with id " + request.getCourseId().toString()));
-        List<Quizz> quizzByCourse = quizzRepository.findByCourse(course);
+                () -> new NotFoundException("Course not found with id " + request.getCourseId().toString()));
 
-        boolean isNameDuplicate = quizzByCourse.stream()
-                .anyMatch(quizz -> quizz.getName().equals(request.getName()));
+        if (!course.getUser().getId().equals(userDetails.getUsername())) {
+            throw new AccessDeniedException("You do not have permission to create a quiz for this course.");
+        }
 
-        if (isNameDuplicate)
-            throw new RuntimeException("Quizz already exists in this course.");
+        if (quizzRepository.existsByCourse(course)) {
+            throw new BadRequestException("Quizz already exists for this course.");
+        }
 
-        ModelMapper modelMapper = new ModelMapper();
         modelMapper.getConfiguration().setSkipNullEnabled(true);
 
         Quizz quizz = modelMapper.map(request, Quizz.class);
+        quizz.setCourse(course);
         quizz.setCreatedDate(new Date());
         quizz.setUpdatedDate(new Date());
 
@@ -54,24 +59,23 @@ public class QuizzService {
     }
 
     @Transactional
-    // create quizz by chapter id
-    public QuizzResponse createQuizzByChapterService(CreateQuizzByChapterRequest request) {
+    public QuizzResponse createQuizzByChapterService(CreateQuizzByChapterRequest request, UserDetails userDetails) {
         Chapter chapter = chapterRepository.findById(request.getChapterId().toString()).orElseThrow(
-                () -> new RuntimeException("Chapter not find with id " + request.getChapterId().toString()));
-        List<Quizz> quizzByChapter = quizzRepository.findByChapter(chapter);
-        Course couseByChapter = chapter.getCourse();
+                () -> new NotFoundException("Chapter not found with id " + request.getChapterId().toString()));
+        Course course = chapter.getCourse();
 
-        boolean isNameDuplicate = quizzByChapter.stream()
-                .anyMatch(quizz -> quizz.getName().equals(request.getName()));
+        if (!course.getUser().getId().equals(userDetails.getUsername())) {
+            throw new AccessDeniedException("You do not have permission to create a quiz for this chapter.");
+        }
 
-        if (isNameDuplicate)
-            throw new RuntimeException("Quizz already exists in this chapter.");
+        if (quizzRepository.existsByChapter(chapter)) {
+            throw new BadRequestException("Quizz already exists for this chapter.");
+        }
 
-        ModelMapper modelMapper = new ModelMapper();
         modelMapper.getConfiguration().setSkipNullEnabled(true);
 
         Quizz quizz = modelMapper.map(request, Quizz.class);
-        quizz.setCourse(couseByChapter);
+        quizz.setChapter(chapter);
         quizz.setCreatedDate(new Date());
         quizz.setUpdatedDate(new Date());
 
@@ -79,48 +83,51 @@ public class QuizzService {
     }
 
     @Transactional
-    // search by course
-    public List<QuizzResponse> getQuizzByCourseService(UUID courseId) {
-        Course course = courseRepository.findById(courseId.toString())
-                .orElseThrow(() -> new RuntimeException("Course not found with id " + courseId.toString()));
-        List<Quizz> quizzes = quizzRepository.findByCourse(course);
-
-        return new ModelMapper().map(quizzes, new TypeToken<List<QuizzResponse>>() {
-        }.getType());
-    }
-
-    @Transactional
-    // search by chapter
-    public List<QuizzResponse> getQuizzByChapterService(UUID chapterId) {
-        Chapter chapter = chapterRepository.findById(chapterId.toString())
-                .orElseThrow(() -> new RuntimeException("Chapter not found with id " + chapterId.toString()));
-        List<Quizz> quizzes = quizzRepository.findByChapter(chapter);
-
-        return new ModelMapper().map(quizzes, new TypeToken<List<QuizzResponse>>() {
-        }.getType());
-    }
-
-    @Transactional
-    // update quizz
-    public QuizzResponse updateQuizzService(UpdateQuizzRequest request) {
+    public QuizzResponse updateQuizzService(UpdateQuizzRequest request, UserDetails userDetails) {
         Quizz quizz = quizzRepository.findById(request.getQuizzId().toString())
-                .orElseThrow(() -> new RuntimeException("Course not found with id " + request.getQuizzId().toString()));
+                .orElseThrow(() -> new NotFoundException("Quizz not found with id " + request.getQuizzId().toString()));
 
-        quizz.setUpdatedDate(new Date());
+        if (!quizz.getCourse().getUser().getId().equals(userDetails.getUsername())) {
+            throw new AccessDeniedException("You do not have permission to update this quiz.");
+        }
 
-        ModelMapper modelMapper = new ModelMapper();
-        modelMapper.getConfiguration().setSkipNullEnabled(true);
         modelMapper.map(request, quizz);
+        quizz.setUpdatedDate(new Date());
 
         return modelMapper.map(quizzRepository.save(quizz), QuizzResponse.class);
     }
 
     @Transactional
-    public QuizzResponse deleteQuizzService(UUID quizzId) {
+    public QuizzResponse deleteQuizzService(UUID quizzId, UserDetails userDetails) {
         Quizz quizz = quizzRepository.findById(quizzId.toString()).orElseThrow(
-                () -> new RuntimeException("Quizz not found with id " + quizzId.toString()));
+                () -> new NotFoundException("Quizz not found with id " + quizzId.toString()));
+
+        if (!quizz.getCourse().getUser().getId().equals(userDetails.getUsername())) {
+            throw new AccessDeniedException("You do not have permission to delete this quiz.");
+        }
 
         quizzRepository.delete(quizz);
-        return new ModelMapper().map(quizz, QuizzResponse.class);
+        return modelMapper.map(quizz, QuizzResponse.class);
     }
+
+    @Transactional
+    public List<QuizzResponse> getQuizzByCourseService(UUID courseId) {
+        Course course = courseRepository.findById(courseId.toString())
+                .orElseThrow(() -> new NotFoundException("Course not found with id " + courseId.toString()));
+        List<Quizz> quizzes = quizzRepository.findByCourse(course);
+
+        return modelMapper.map(quizzes, new TypeToken<List<QuizzResponse>>() {
+        }.getType());
+    }
+
+    @Transactional
+    public List<QuizzResponse> getQuizzByChapterService(UUID chapterId) {
+        Chapter chapter = chapterRepository.findById(chapterId.toString())
+                .orElseThrow(() -> new NotFoundException("Chapter not found with id " + chapterId.toString()));
+        List<Quizz> quizzes = quizzRepository.findByChapter(chapter);
+
+        return modelMapper.map(quizzes, new TypeToken<List<QuizzResponse>>() {
+        }.getType());
+    }
+
 }
